Refactoring, removed namespace WinSwag.Models
