+++ WinSwag/ViewModels/ForModels/OperationViewModel.cs
