+++ WinSwag/WinSwag.csproj
@@ -117,8 +117,8 @@
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
     <Compile Include="ApplicationInstance.cs" />
-    <Compile Include="Models\SettingsDictionary.cs" />
-    <Compile Include="Models\SwaggerSessionManager.cs" />
+    <Compile Include="SettingsDictionary.cs" />
+    <Compile Include="Services\SwaggerSessionManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Services\ApplicationInfo.cs" />
     <Compile Include="Services\ISessionManager.cs" />
@@ -141,11 +141,11 @@
       <DependentUpon>ResponseTemplates.xaml</DependentUpon>
     </Compile>
     <Compile Include="ViewModels\DateTimeToStringConverter.cs" />
-    <Compile Include="ViewModels\ForModels\SwaggerOperationViewModel.cs" />
+    <Compile Include="ViewModels\ForModels\OperationViewModel.cs" />
     <Compile Include="ViewModels\OperationManagerVM.cs" />
     <Compile Include="ViewModels\ParameterTemplateSelector.cs" />
     <Compile Include="ViewModels\ResponseTemplateSelector.cs" />
-    <Compile Include="ViewModels\ForModels\SwaggerDocumentViewModel.cs" />
+    <Compile Include="ViewModels\ForModels\OpenApiDocumentViewModel.cs" />
     <Compile Include="ViewModels\SessionManagerVM.cs" />
     <Compile Include="Views\ApiInfoPage.xaml.cs">
       <DependentUpon>ApiInfoPage.xaml</DependentUpon>
